MGMT-8890: Integrate support for multiarch release images (#4189)
* MGMT-8890: Integrate support for multiarch release images

This PR adds support of multiarch OCP release images to the Assisted
Service.

The following happens now when a multiarch release image is provided via
the ClusterImageSet

1) cluster.CPUArchitecture is set as "multiarch"
2) validation of existence of OS image matching version and architecture
   is offloaded to the InfraEnv when the ISO is created
3) validation of matching cluster's and InfraEnv's architecture is
   removed

The rationale for (1) is that if using multiarch release image, the
cluster is no longer bound to one particular architecture but can rather
be deployed with any architecture as long as release image contains an
image for it and OS image that matches the version and architecture
exists.

The rationale for (2) is that the set of architectures supported by the
release image can be bigger than set of architectures supported by OS
images available for the Assisted Service. Given that we cannot enforce
users to provide OS image for every arch from the OCP release image,
this validation happens when InfraEnv tries to create an ISO. If it
happens that InfraEnv is created for an architecture supported by the
release image but without OS image available, a clear error message will
be shown.

The rationale for (3) is that this work is a preliminary work to support
heterogenous OCP clusters. Therefore we don't want to strictly limit
InfraEnvs bound to the specific ACI to contain only single architecture.

Given that (2) and (3) are technically allowing for some installation
schemas that we may not want to immediately support, we may want to add
additional validations in a subsequent PR. This is however left outside
of the scope of this change here as it clearly depends on the timelines
for releasing specific features like multiarch OCP images and
heterogenous OCP clusters.

Supersedes: #4185
Contributes-to: MGMT-8890
Closes: MGMT-11353
Closes: MGMT-11354

* MGMT-8890: Add debug logging for release image handler

With this commit we want to print additional debug information regarding
the process of detection version and architecture of provided OCP
release image. It is going to be helpful when handling multiarch release
images.