MGMT-11447: Improve ignition downloadable (day2 API connectivity check) validation message (#4237)
The current message (constructed by the UI) is "API IP connectivity
failure - To continue installation, configure your DNS or alternatively
<a>Update cluster with API IP</a>"

There are multiple issues with the current message:

1) The title says "API IP connectivity failure". This is not technically
   correct. The issue is reaching the API in general. It doesn't
   necessarily have anything to do with an "IP address".

2) The suggestion to "configure your DNS or Update cluster with API IP"
   is misleading - it seems to imply that if the user chooses to not
   configure their DNS, they can simply "update the cluster with API IP"
   and everything will work fine. This is not necessarily correct -
   none-platform clusters require the user to configure DNS, regardless
   of whether they provide us with the IP address of their cluster or
   not. The worker will simply not be able to join without it.

3) "update the cluster with API IP" needs rewording. "Updating a
   cluster" doesn't mean much to someone without internal AI developer
   perspective. It's extra confusing in day-2 because they already have
   a cluster. The concept of a "cluster" that the message is referring
   to is unknown to the user - us reusing the day-1 internal "cluster"
   entity for adding day-2 workers is merely an implementation detail
   and not something they user should be aware of, yet this message
   expects them to be aware of it. When read superficially by a user,
   this message seems to imply they have to make some change to their
   actual cluster, and it's confusing, because they don't.

To solve this we're gonna make the backend build most of the message.

1) The title should instead say "API connectivity failure"
- This will be done as part of the UI PR

2) The message should be more verbose about what actually happened:
   "This host has failed to contact the cluster's API at <API URL> with
   the following error: <error>. Please ensure the host can reach the
   API", and to that the UI will add to that something like
   "Alternatively, set a different API hostname <link>" which will allow
   lead the users to change the hostname setting dialog

It's important to include the entire URL in the message because we're
trying to contact the API at a very specific port number, and this is a
port number that's likely to be blocked by a firewall, so they user
should be aware of it. It's also important to include the entire URL
because the API domain name is something the UI crafts without user
interaction, when importing the cluster from OCM, so it's possible that
it's not even correct - the user needs to see it.

The agent will be separately improved to provide the service with both
the URL it tried to contact (which it itself will get from the service,
but we include it in the response anyway in case it changed and the
agent didn't get a chance to try the new URL yet) and the agent will
also include the error message it encountered when attempting to
download the ignition from the API - DNS error, connection refused,
whatever - the user should not be forced to SSH and watch agent logs to
figure out what went wrong.

Also changed the validation to be a `ValidationError` whenever the
service fails to deserialize the agent response, instead of a
"ValidationFailure".