MGMT-8890: Extend release image struct to store all archs
This PR extends the ReleaseImage with a new `[]string` field that stores
all the architectures referenced by the release image. This is useful
for working with multiarch images when `releaseImage.CPUArchitecture`
field has only "multiarch" as a value but it carries no knowledge about
specific architectures of such an image.

With this change a multiarch release image will be stored as e.g.

```
CPUArchitecture: "multiarch"
CPUArchitectures: ["x86_64", "ppc64le", "s390x", "arm64"]
```

In order to allow for further simplifications in the future and
potentially a drop of singular `CPUArchitecture` field, the value of
plural field for a single-arch image will be a one-element list, e.g.

```
CPUArchitecture: "x86_64"
CPUArchitectures: ["x86_64"]
```

Contributes-to: MGMT-8890
Closes: MGMT-11494