MGMT-10789: Make sure that service can connect to KubeAPI during (#4135)
Day2 installation

Failures have been witnessed during Day2 installation that have been
caused by the absence of a valid KubeAPI connection from the service to
the spoke cluster.

This validation will execute whenever we add a host to a Day2 cluster to
ensure that a check is carried out as early as possible so that the user
can rectify the problem with appropriate DNS entries etc.