MGMT-11156: Cluster validation, merge the print and validate methods (#4130)
* MGMT-11156: Cluster validation, merge the print and validate methods

Same as 4e04230 but for cluster validations instead of host validations

The present implementation of cluster validation has a problem, it splits the validation into two phases. The validation and the printing of the validation message.

This leads to a situation in some cases where it is not possible to use the output of a validation to determine the message that should be sent back to the user, there are some workarounds but these involve poor practices such as code repetition.

This commit changes the validation to be a single method. Due to the complexity of the refactor, no effort has been made to make each of the validation methods their "most efficient" it is assumed that these fixes will take place in subsequent commits.

Because the validation methods made heavy use of return statements, the style has had to be changed to an if/else style and some code has had to be reorganised a little.

In theory, the existing tests should be sufficient to act as a "smoke" test for this functionality provided that there is enough coverage.

* MGMT-11156: Adding tests for isNetworksSameAddressFamilies in validator