MGMT-11543: When there are many connectivity check packet losses, the host monitoring thread gets stuck (#4259)
When there are many connectivity check packet losses, the packet loss validation
status message contains list of the failed hosts, and the failure
percentage for each.
In order to create the list of failed hosts, each failed host is
identified by its IP address.  Since the IP address is in the inventory,
to identify the host all inventories in the cluster need to be
unmarshalled.
The change is to identify the host by host-id and not by the IP address.
In this case, inventory unmarshalling is not needed for host
identification, because the search is by host id.