MGMT-11542: Fix connectivity check running forever on large cluster (#4253)
When having a large cluster that has many varying connectivity problems,
there are many possible groups.  Finding the optimal one takes a very
long time. This causes cluster monitor to never return which causes the
assisted installer pod to be restarted due to health problems.
The fix to this problem, is to provide a semi optimal problem. In case
there is full mesh connectivity between all nodes, the change will
provide the correct answer which means full connectivity.
In case that there are many possible full mesh connectivity groups, it
might not find the optimal one  (although in the tests that were run the
previous algorithm and the current one produce same results).