Bug-2107617:  Attempted 240 node cluster with MCE / Assisted-installer and ended up with 187 node cluster (#4265)
In order to move cluster to finalizing, 3 masters and 2 workers have to
be installed.  The others are still may not be installed yet.
If the cluster installation is complete (operators are up and kubeconfig
in place) but some nodes have not completed installation, these hosts
will never move to installed because they depend on the controller to
report the stage.
The controller stops running if cluster is not in installation status.
This means that these hosts will eventually move to error on timeout.
This change ensures that cluster can move from finalizing to insllled
only if all its hosts are either installed, or in error.
This change also enables transition from finalizing to
pending-user-action if there is a host in pending-user-action while the
cluster is finalizing. The cluster will move back to finalizing when the
host moves back to installing status.