Bug 2116549: SNO dnsmasq should use the managed nm resolv.conf
# Background

NetworkManager stores its generated `resolv.conf` configuration in
`/run/NetworkManager/resolv.conf`. Since we use the `rc-manager=unmanaged`
config, it means NetworkManager only generates the above file, but
doesn't copy it to `/etc/resolv.conf`.

We use a network-manager dispatcher script we call `forcedns` to copy
`/run/NetworkManager/resolv.conf` to `/etc/resolv.conf`, but before we
do that, we also append the node's IP address to that file as a
nameserver. This is done to make the pods on the node use the dnsmasq
DNS server as their DNS server, and this way they can resolve the
API/Internal API/Ingress domains that we hard-code into the dnsmasq's
configuration.

# Issue

There's an OCP/RHCOS 4.11 bug that's still under investigation where
dnsmasq's CPU usage gets stuck at 100%. We have reasons to believe it
has something to do with us modifying/moving the `/etc/resolv.conf` file
while dnsmasq is using it. Also we suspect newer versions of `dnsmasq`
may not like the node's own IP address appearing in the
`/etc/resolv.conf` file.

# Proposed solution

Configure dnsmasq to use `/run/NetworkManager/resolv.conf` as its
`resolv-file` instead of `/etc/resolv.conf`. This way it uses a file
that's stable and doesn't change, and also it uses a file that doesn't
contain the node's own IP address.

# Note

The dnsmasq bug is under investigation and would have to be fixed in
dnsmasq itself in order to fix existing clusters that may want to
upgrade to 4.11. This PR / solution would alleviate the issue for newly
installed Assisted clusters.