Bug 2113954: Fix order when merging ignitions (#4240)
This PR changes the order of calling ignition merge function in BMAC.
This is because the function is not performing a merge towards the
latest provided ignition but merges towards the ignition version
provided as a second parameter.

Because of this behaviour we have observed cases where user providing
ignition override v3.2.0 was receiving a rendered ignition v3.1.0 with
some data lost (i.e. features introduced in 3.2 and not available in
3.1).

Fixes: Bug-2113954
Fixes: Bug-2099929
Closes: MGMTBUGSM-447
Closes: MGMTBUGSM-499

Co-authored-by: Mat Kowalski <mko@redhat.com>