MGMT-11736: Support ASC in multi namespaces
To support deploying an assisted infrastructure operator on zero workers hyper shift we need first to support deploying the operator on multiple namespaces.

To support backward compatibility, if the operator is deployed to the assisted-installer namespace we assume that the setup is a cluster scoped deployment. Otherwise, we implicitly assume that the hyper shift mode is on.