MGMT-8890: Use "multi" as a keyword for multi-arch OpenShift (#4293)
This PR changes the value used to mark multiarchitecture OpenShift
clusters from "multiarch" to "multi". This is because the official
version marker is `-multi` so that all the tooling relying on this name
does not have to perform any translation.

An example of the currently performed translation is "amd64" and
"x86_64" that are representing the same architecture in two different
ways. With this change here we will not need to handle "multi" and "multiarch"
as the only used one will be "multi".

The change is safe as there are no clusters created using multiarch
release payload yet.

Contributes-to: MGMT-8890