MGMT-10850: Message for Day2 reboot is misleading (#4298)
The install progress message in for a node is misleading when a Day2
host is rebooting as it states that "No further updates will be posted"

This is inaccurate as there are certain activities that could lead to an
update to the agent status. For example CSR auto signature operations.

This commit rectifies this problem by removing the confusing part of the
message.