MGMT-8890: InfraEnv to validate content of release image (#4285)
This PR adds a logic that ensures we create InfraEnvs that refer to
existing release images. This is in order to prevent the following
scenario

* release image contains architecture [x86_64 ppc64le s390x]
* OS images available for [x86_64 arm64]
* infraenv created for arm64

The current logic for creating the ISO already detects the case when
InfraEnv uses architecture for which no RHCOS image is available.
However a scenario where RHCOS exists but OpenShift release image does
not was not covered till now. Without the change introduced by this PR
the missing release image would be detected only during the
installation. With this change it will happen very early in the
process so that we can prevent the installation from even starting.

Closes: MGMT-11500
Closes: MGMT-11569