MGMT-8890: Update openshift-versions API call to return all archs (#4297)
This PR modifies the openshift-versions API call so that for multiarch
release images it unpacks the supported architectures. This is required
so that instead of response containing

```
"4.11":{"cpu_architectures":["multi"]
```

we get a list of architectures from this image. With the current
behaviour it is effectively impossible to query the service about
possible architectures that are supported. This PR removes the
limitation.

Contributes-to: MGMT-8890
Closes: MGMT-11758