MGMT-8890: Add OS image validator to openshift-versions API call
With this commit we are changing behaviour of the "openshift-versions"
API call so that it returns only versions and architectures if there is
a matching release image as well as RHCOS image.

This is in order to cover a scenario when a multi-arch release image is
used by the service, but available OS images do not contain every
architecture that is provided by the release. For such a scenario we
only want to mark availability for those combinations that have a
coverage in both release as well as RHCOS image.

Contributes-to: MGMT-8890
Closes: MGMT-11780