MGMT-8890: Mark cluster as multi-arch automatically (#4314)
This PR introduces a change that marks all the cluster using multiarch
release image as multiarch clusters. This is in order not to force users
(whether via API or via UI) to manually set that property if based on
the selection of the release image we know that the cluster is able to
consist of multiple architectures.

Contributes-to: MGMT-8890