MGMTBUGSM-482: kube-api - improve error msg for missing OS image
When creating a ClusterDeployment we call AddReleaseImage to
ensure the relevant release image is available.
If an applicable OS image is missing, the sync condition should
include a proper message. Hence, added to the underlined issue
to the condition message.

E.g.
The Spec could not be synced due to backend error: failed to add release image 'quay.io/openshift-release-dev/ocp-release:4.7.52-x86_64'. Please ensure the releaseImage field in ClusterImageSet 'openshift-v4.11' is valid (error: No OS images are available for version 4.7.52 and architecture x86_64).