MGMT-11739: added KubeconfigSecretRef property to ASC
In order to communicate with an external control-plane
(e.g. hypershift HostedCluster), the service should be
supplied with the relevant kubeconfig.
I.e. When applying the service deployment in AgentServiceConfigReconciler,
a path for a custom kubeconfig should be supplied in an env var.

Hence, introduced the following changes, added KubeconfigSecretRef
property to AgentServiceConfigSpec. This property is a reference
for a secret that contains the kubeconfig for an external cluster.

The logic for adding relevant Volume, VolumeMount and EnvVar for the
service deployment will be added as part of: #4322