Start reclaim agent as a DaemonSet
This is required becase the pod needs to be started on a node that is
cordoned and has been drained. This also means that we need a service
account and cluster role binding to ensure the pod runs with the
privileged security context.

Additionally the daemon set is marked as owned by the node so it is
garbage collected when the node is eventually deleted. This prevents the
pod from running if the node is subsequently added back to the cluster
with the same hostname.