MGMT-8568: Use NMState GenerateConfiguration Go binding (#3870)
This change replaces the code nmstate configuration validation from
using a binary named nmstatectl to a dynamic nmstate Go binding.

Using the Go binding, the code makes use of the GenerateConfiguration function
to validate nmstate configurations, which is a lot faster and consumes
less resources compared to a use of an external binary.

In addition, changes to how we compile the code were made to support
the above-mentioned Go binding.