MGMT-11888: Pass usable NIC's MAC address to install-config
Currently in order to fill the BootMACAddress in install-config.yaml we
are picking any NIC that has non-zero MAC address and use this value.
It's been proven that this logic is prone to fail for servers that have
NICs with duplicated MACs even though those NICs are disabled and not
used.

This PR changes the logic so that we will only pick MAC address of a NIC
if it has an IP address assigned. Moreover, the IP address needs to
belong to the machine network of the cluster.

Closes: MGMT-11888