NO-ISSUE: Generate vendor only after generating swagger
# Background

Our `api/` and `models/` directories are both separate go modules
generated by `generate-from-swagger`.

Our repo root module depends on both of them, and `api/` depends on
`/models`.

Today when we run `make generate` we first run `go mod vendor` inside
all modules and only then we re-generate the `api/` and `models/`
modules from swagger.

# Issue

Doing it in that order means that the vendoring performed in the root
module and inside the `api/` module, which both depend on `models/`,
will initially be incorrect, as it will perform the vendoring with an
outdated `models/` directory. For now users could solve this by simply
running `make generate` again, but that was slow, confusing, and
annoying.

# Solution

Run `generate-from-swagger` before `generate-vendor` so that vendoring
always uses the fresh `models/` rather than the older `models/` that
does not contain potential swagger changes.