Update json-iterator and reflect2 dependencies to fix panic
When reading and then subsequently writing a BMH yaml within a bootstrap
ignition file we encountered a panic.

This seems to be a rather well known issue with these two module
versions and upgrading fixes the problem.

This was fixed in master by cc8f3538 when updating some seemingly
unrelated packages, but it seems safer to just updated the required ones
rather than also bringing back the prometheus client update.

References:
https://github.com/golang/go/issues/48238
https://github.com/kubernetes/kubernetes/issues/104947

https://bugzilla.redhat.com/show_bug.cgi?id=2122747