MGMT-11563: Explicitly mark multiarch release payloads (#4373)
With this commit we want to specifically mark multiarch release payloads
with "-multi" suffix. This is in order for customers to be able to
explicitly request the installation using such an image without having
to deal with the `cluster.CPUArchitecture` property.

With this change the selection can be easily done via the existing UI as
opposed to requesting multiarch payload via `cluster.CPUArchitecture`
which currently has no UI support.

Contributes-to: MGMT-11563
Contributes-to: MGMT-11859