MGMT-11563: Add Z-stream to multi-arch release payload
This patch adds Z-stream to the OCP version for the multiarch release
payload. This is because when creating a cluster, the object is
backfilled with the openshift_version value coming from the full version
on the release payload, i.e. when requesting a cluster with "4.11-multi"
we will backfill the cluster with value of "4.11.0-multi" as the value
configured in the release image.

This is confiusing and creates issues when parsing available release
images afterwards, as the match between `4.11.0-multi` and `4.11-multi`
is non-obvious and would require carrying the suffix around all the time
whenever we strip Z-stream.

The other solution would be to collapse multiarch payload to simple
`4.11` but this has already proven in the past to be potentially
confusing as the same version would be availalbe for single- and
multiarch payload.

Contributes-to: MGMT-11859
Contributes-to: MGMT-11563