MGMT-8987: Allow for one worker to be added when masters are schedulable (#3239)
* MGMT-8987: Allow for one worker to be added when masters are schedulable

This commit changes the masters validation so that adding one worker as part of the cluster deployment
is allowed iff the masters are schedulable. This architecture is being used for other workflows like ZTP
and factory pipelines. It's also been reviewed by OCP experts.

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* MGMT-8987: Allow for creating clusters with one worker

After some discussion with the OCP team in this enhancement[0], it came out that it is ok
to allow for 3+1 clusters to be created. This does not mean the architecture is supported but that
it is possible to create such clusters. IPI does not prevent users to do this and, with the right
Infrastructure CR configuration, it is possible to create a healthy, HA, cluster with 3+1.

This is a first step towards adding to AI the capability to deploy healthy 3+1 clusters. More commits to
come

[0] https://github.com/openshift/enhancements/pull/1057

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* MGMT-8987: Set InfrastructureTopology as HA for 4-nodes

This commit overwrites the InfrastructureTopology in the Infrastructure
CR manifest when a 4-nodes cluster is requested. It odes it using a
`yaml-patch` format and applying it on top of the already-generated
infrastructure manifest.

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* Add yaml-patch vendor

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* Assume 4 nodes means schedulable masters

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* Add debug messages for the infrastructure patch step

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* Update unit-tests

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* update comment explaining min required workers

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* Address review comments

Signed-off-by: Flavio Percoco <flavio@redhat.com>

Signed-off-by: Flavio Percoco <flavio@redhat.com>
Co-authored-by: Flavio Percoco <flavio@redhat.com>