MGMT-10968: Use a daemon set for the spoke cluster agent during node reclaim (#4348)
* Update openshift api version

This is required to get access to ClusterRoleBinding types

* Remove varnamelen from go.sum

This was causing a build failure and `go mod why` reports that it isn't
required so I removed it and things seem fine

* Add authorization scheme to spoke kube client

This scheme contains ClusterRoleBindings which will be used to associate
the reclaim service account with the privileged cluster role.

* Start reclaim agent as a DaemonSet

This is required becase the pod needs to be started on a node that is
cordoned and has been drained. This also means that we need a service
account and cluster role binding to ensure the pod runs with the
privileged security context.

Additionally the daemon set is marked as owned by the node so it is
garbage collected when the node is eventually deleted. This prevents the
pod from running if the node is subsequently added back to the cluster
with the same hostname.

* Use controllerutil.CreateOrUpdate when creating spoke resources

It's reasonable to assume that the assisted-installer namespace could be
present when we go to reclaim a node. This will allow us to more easily
update that namespace with the labels required by kubernetes Pod
Security Standards for running privileged pods.

Additionally this is a nice way of deduplicating the NotFound error
handling.

* Add pod security labels to assisted-installer namespace for reclaim

These will ensure that we're able to run privileged pods in this
namespace when Pod Security Standards are adopted in OCP 4.12

* Use Role and RoleBinding rather than ClusterRoleBinding

This keeps everything required for reclaim in the namespace rather than
creating cluster-scoped resources

* Log the result of CreateOrUpdate when it does something