Remove host reclaim feature flag (#4398)
This was needed so that we could merge the work a bit at a time.
Now that all the pieces are in place for reclaim to work correctly this
can be removed.

This isn't a problem for non-kubeapi deployments as the agent controller
is the only place we will ever run the reclaim transition.

https://issues.redhat.com/browse/MGMT-12062