OCPBUGS-1482: Don't override schedulableMasters unnecessarily
When generating the Ignition files, the installer already sets
schdulableMasters to true when there are no worker nodes (i.e. in the
SNO and compact cluster topologies. (See
https://github.com/openshift/installer/pull/2004) Therefore it is
unnecessary to override it here (though it may be preferred to avoid a
warning log from the installer).

Since https://github.com/openshift/installer/pull/6247, attempting to
override the schedulableMasters setting causes installation to fail,
because there are two manifests of the same type and name that conflict.

Since we don't need to set this override when the installer would
already do it, avoid doing so and triggering the error when the value is
determined by the number of hosts rather than explicitly set by the
user.

The conflict still needs to be resolved so that the user can enable
schedulableMasters, but this at least allows the SNO and compact
topologies to install OpenShift 4.12 again.

This partially reverts commit c45f3699d83665a5e27a00939a9ff40907f39111.