MGMT-12117: fix flaky infra-env UT
on a fast machine there is chance to get a case where infra env is created at
the same time that the GC function try to delete it.
It can cause a case where the infra env is not deleted and the tests fails.
Adding several seconds to the GC function will make sure that the interval
will see the infra-env as old

5 sec is random number, it could be any amount of time that will be
  different then the creation time of the infra-env
I decided to add the time to the GC function to void channging the
default behavior of the db that set the creation time