AGENT-271: Make generation of kubeadmin-password optional
In the automated flow of the ephemeral installer, we generate the
kubeadmin password at the same time as the ISO. To avoid having to store
the password in the ISO, we pass only the hash. When running the
installer to create the bootstrap ignition, it reads this hash and
therefore does not create a kubeadmin-password file.

To allow this, do not fail when the kubeadmin-password file does not
exist.