MGMT-8460: Automatically rename discovered hosts
Currently host that use a forbidden name like `hostname` need to be
manually renamed by users before proceeding to the installation. That
isn't convenient for users. To avoid it this patch changes the code so
that when the inventory is updated those hosts will be automatically
renamed with a name calculated from the MAC address a network
interface card of the host that has a non-local IP address.

Related: https://issues.redhat.com/browse/MGMT-8460
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>