Fix panic when infraEnv image has not been created
Previously the preprovisioning controller was unconditionally looking
for the created time in the infraEnv status. This would cause the entire
assisted-service container to crash if the created time was not set.

This is only set after the infraEnv controller sets the image URLs and
may not be set for some time if there are, for example, format errors in
nmstate config associated with the infraEnv. Once the container starts
crashing it is unlikely to recover even if the underlying issues are
resolved as the preprovisioning controller seemed to consistently reconcile
before the infraenv controller got a chance to set the created time.

This patch checks that the image created time is set and sets a
condition if not. The preprovisionsing image controller is watching for
infraenv events so it will reconcile when the image gets created.

https://issues.redhat.com/browse/MGMT-12188