MGMT-8890: Allow weak matching for GetReleaseImage
This PR extends the GetReleaseImage function to allow for weak matching
between requested and available OpenShift versions. This is to allow
queries for e.g. `4.12` when only `4.12.1` is available.

The extensions allows to use the GetReleaseImage for validation purposes
when the caller only cares about a non-empty output, but does not care
which exactly release image is matched.

This has been discovered only after multi-arch clusters got the
validation between InfraEnv (i.e. RHCOS image) and OCP release. With the
ability to perform weak matching we don't need to align RHCOS image to
the z-stream of OCP, so e.g. we can offer `4.12.0-ec.3-multi` to be
served by `4.12` RHCOS (without that change we need to label RHCOS
explicitly as `4.12.0-ec.3-multi`).

Single-arch clusters are not affected, as for such a cluster InfraEnv
does not perform any validation whatsoever.

Closes: Bug-2131396
Closes: MGMTBUGSM-567
Caused-by: MGMT-11500
Caused-by: #4285
Contributes-to: MGMT-8890