Update infraEnv initrd URL when image content is updated (#4461)
Previously only the ISO URL was getting updated when the image content
changed. This will be a problem when assisted service is configured to
boot machines through ironic using the initrd. In this case when
something is changed in the infraenv that should change the content of
the initrd the infraenv, and thus, the preprovisioning image were not
being updated.

With this change the infraenv controller will recreate the token on the
initrd URL whenever the ISO URL changes. This will force ironic to
redownload the image and cause the machines to reboot, applying the
changes.

https://issues.redhat.com/browse/MGMT-8864