MGMT-8890: Create validator for RHCOS and OCP versions
This PR creates the `ValidateReleaseImageForRHCOS` function responsible
for validating whether a specified RHCOS version has a matching OCP
release available, as otherwise it makes no sense to create such an
InfraEnv as it would be effectively unusable until a valid OCP version
appears.

This has been discovered only after multi-arch clusters got the
validation between InfraEnv (i.e. RHCOS image) and OCP release. With the
ability to perform weak matching we don't need to align RHCOS image to
the z-stream of OCP, so e.g. we can offer `4.12.0-ec.3-multi` to be
served by `4.12` RHCOS (without that change we need to label RHCOS
explicitly as `4.12.0-ec.3-multi`).

Single-arch clusters are not affected, as for such a cluster InfraEnv
does not perform any validation whatsoever.

Closes: Bug-2131396
Closes: MGMTBUGSM-567
Caused-by: MGMT-11500
Caused-by: #4285
Contributes-to: MGMT-8890
Supersedes: #4464