Bug 2106110: Don't set ip=<nic>:dhcp kernel params with static networking (#4475)
This PR changes the behaviour of logic responsible for appending
`ip=<nic>:dhcp` kernel parameters. Those must not be added when static
network configuration is used as in some specific cases it causes loss
of connectivity due to NetworkManager having to deal with two
contradicting configurations (static config created by NMstate and DHCP
as requested by kernel parameter).

With this change as soon as we detect that InfraEnv has network
configuration provided by the user, we will not add any `ip` kernel
param. This is not expected to cause any trouble as the behaviour will
be as follows

* with no static configuration, `ip=<nic>:dhcp` will be set on the NIC
  that is used to provision the server (behaviour like today)
* with static configuration the NIC used to provision the server should
  have IP address assigned with NMstate

Closes: Bug-2106110
Closes: MGMTBUGSM-508