MGMT-11129: Only set bootstrap host when it's attached to a cluster (#4483)
https://issues.redhat.com/browse/MGMT-11129
Initially there wasn't a check to see if the host was attached to a
cluster before setting the bootstrap host. In a scenario like
late-binding, there could be hosts without a cluster, which would
cause a panic.

This checks to make sure the cluster is not nil before executing
the logic to set a bootstrap host.