Bug 2107715: Add validation for Agent's role
This PR adds validation for the Agent's role whenever an update is
requested. This is in order to prevent configuration where an incorrect
role causes the Node to try to download a non-existing ignition file and
fail the installation without a clear error returned to the end user.

With this change, as soon as an attempt to set an invalid role is
detected, an API error will be returned preventing the change to be
applied (in the REST API flow) or causing Agent CR to return a SpecSync
error (in the Kube API flow).

Closes: MGMTBUGSM-473