MGMT-12261: Respect `InstallConfigOverrides` when `getCAContents` is non-empty
# Background

Users can influence the install-config `AdditionalTrustBundle` through `InstallConfigOverrides`.

This generally works. Except for in the situation described below.

# Issue

`getCAContents` is a function that in some circumstances returns
internal CA certificates that need to be set in the install-config
`AdditionalTrustBundle`.

In most cases, `getCAContents` returns an empty string and everything
works fine.

But when `getCAContents` actually returns certificates, then we
completely overwrite the install-config `AdditionalTrustBundle`,
even when the user modified it through `InstallConfigOverrides`.

This is bad.

The situations in which `getCAContents` is non-empty is when the
Assisted Service admin deploys it with either:

- Red Hat Root CA (`INSTALL_RH_CA` env)
    - This is the case in our staging environment and this is the cause
      for the issues seen in MGMT-12261
- Mirror registry configuration mounts, as detected by our mirrorregistries package
    - I'm not aware of anyone that ran into this particular issue, but
      it theoretically could happen

# Solution

We should instead merge the `InstallConfigOverrides` certs with the certs we
receive from `getCAContents`, rather than overwrite them.